desktop-app/requirements.txt
python-dotenv
requests
keyring
google-auth
google-cloud-firestore
cowsay